refactor(probe): migrate dns planning into per-protocol plugin
